Разделил обработ. по тематике в отдел. файлах. Отдельно инициализ. бота.
