+++ handlers/__init__.py
@@ -1,0 +1,3 @@
+from handlers import handler_commands
+from handlers import handler_send_photo
+from handlers import handler_just_text
