+++ init_bot.py
@@ -1,0 +1,6 @@
+from aiogram import Dispatcher, Bot
+from settings import config
+
+
+bot = Bot(token=config.API_TELEGRAM_TOKEN)
+dispatcher = Dispatcher(bot=bot)
