+++ pyproject.toml
@@ -11,6 +11,7 @@ python = "^3.11"
 python-telegram-bot = "^13.15"
 aiogram = "^2.23.1"
 mypy = "^0.991"
+loguru = "^0.6.0"
 
 
 [build-system]
